requirements.txt
pillow
pytesseract
pyperclip
pyautogui
pygetwindow
tk
